move DisposableHandler to thirtyinch-rx2 module
